\name{mbiedata-package}
\alias{mbiedata-package}
\alias{mbiedata}
\docType{package}
\title{
Data from the New Zealand Ministry of Business, Innovation and Employment
}
\description{
Data, mostly on tourism, produced by the NZ Ministry of Business and Employment that is either slowly changing (once per year) or from a finished data collection.}
\details{
\tabular{ll}{
Package: \tab mbiedata\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-1-31\cr
License: \tab GPL-3\cr
}
This package provides the following rectangular datasets, accessible by \code{data(...)} and with help available from ?.  These are datasets that are either finished collection (RVM, the old IVS methodology files) or only infrequently updated (RTEs, forecasts)
\enumerate{
  \item{IVStrips} - International Visitor Survey, "old method" 1997-2013
  \item{IVSspend}
  \item{IVSvisits}
  \item{RTEs} - Regional Tourism Estimates
  \item{forecasts} - New Zealand tourism forecasts
  \item{DTSresp} - Domestic Travel Survey
  \item{DTStrips}
  \item{DTSvisits}
  \item{RVM} - Regional Visitor Monitor
  }
  

}
\author{
MBIE Sector Performance team

Maintainer: Peter Ellis <peter.ellis@mbie.govt.nz>

}
\references{
www.mbie.govt.nz
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%%~~ documentation directory ~~



\note{
Crown Copyright. This data and software is protected by copyright. Unless indicated otherwise (either below or within specific items or collections), this copyright material is licensed for re-use under the Creative Commons Attribution 3.0 New Zealand licence. In essence, you are free to copy, distribute and adapt the material, as long as you attribute it to the Ministry of Business, Innovation and Employment and abide by the other licence terms. Please note that this licence does not apply to any logos, emblems and trade marks or to website or email design elements or to any photography. Those specific items may not be re-used without express permission.
}


\keyword{ package }
\seealso{

\code{\link[<pkg>:<pkg>-package]{mbie}}, \code{\link[<pkg>:<pkg>-package]{mbiemaps}}
}
\examples{


}
